wish

+++ src/services/author.service.ts
@@ -1,28 +1,29 @@
 import { FilterQuery } from 'mongoose';
 import {authorModel} from '../models/author';
 import { Iauthor } from '../types/model.types';
+import { userModel } from '../models/user';
 
 export const createAuthor=async (input:FilterQuery<Iauthor>)=>{
-    const author=await new authorModel(input).save();
+    const author=await userModel.create({...input,role: 'author'});
     return author;
 }
 
 export const findAuthor=async (query:FilterQuery<Iauthor>)=>{
-    const author = await authorModel.findOne(query);
+    const author = await userModel.findOne({...query,role: 'author'});
     return author;
 }
 
 export const findAllAuthor=async()=>{
-    const authors=await authorModel.find();
+    const authors=await userModel.find();
     return authors;
 }
 
 export const updateAuthor=async(query:FilterQuery<Iauthor>,update:Partial<Iauthor>)=>{
-    const author = await authorModel.updateOne(query,update);
+    const author = await userModel.updateOne({...query,role:'author'},update);
     return author;
 }
 
 export const deleteAuthor=async (query:FilterQuery<Iauthor>)=>{
-    const author = await authorModel.deleteOne(query);
+    const author = await authorModel.deleteOne({...query,role:'author'});
     return author;
 }
